refactor: fix typo `comfirm` -> `confirm`
