+++ workspaces/client/app/routes/dashboard/admin/certifications/internal/components/certification-table.tsx
@@ -2,7 +2,7 @@ import { useActionState, useCallback, useState } from "react";
 import { Check, Edit, Trash } from "react-feather";
 import { css } from "styled-system/css";
 import { DeleteConfirmation } from "~/components/feature/delete-confirmation";
-import { ConfirmDialog } from "~/components/logic/callable/comfirm";
+import { ConfirmDialog } from "~/components/logic/callable/confirm";
 import { Form } from "~/components/ui/form";
 import { IconButton } from "~/components/ui/icon-button";
 import { Table } from "~/components/ui/table";
