+++ workspaces/client/app/routes/dashboard/admin/equipments/internal/components/equipments-editor.tsx
@@ -3,7 +3,7 @@ import { Edit, Plus, Trash } from "react-feather";
 import { css } from "styled-system/css";
 import { DeleteConfirmation } from "~/components/feature/delete-confirmation";
 import { UserDisplay } from "~/components/feature/user/user-display";
-import { ConfirmDialog } from "~/components/logic/callable/comfirm";
+import { ConfirmDialog } from "~/components/logic/callable/confirm";
 import { ButtonLike } from "~/components/ui/button-like";
 import { IconButton } from "~/components/ui/icon-button";
 import { Table } from "~/components/ui/table";
